pyproject.toml
[tool.ruff]
ignore = ["E501"]

[tool.poetry]
name = "healthfirstai-prototype"
version = "0.1.0"
description = ""
authors = ["Lawrence Lim <ll4715@nyu.edu>"]
readme = "README.md"
packages = [{include = "healthfirstai_prototype"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.220"
cohere = "^4.11.2"
pytest = "^7.4.0"
openai = "^0.27.8"
unstructured = "^0.7.12"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.0"
faiss-cpu = "^1.7.4"
pymysql = "^1.1.0"
fastapi = "^0.100.0"
redis = "^4.6.0"
sqlalchemy = "^2.0.18"
cryptography = "^41.0.2"
pgvector = "^0.1.8"
psycopg2-binary = "^2.9.6"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.3"
selenium = "^4.10.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = 'healthfirstai_prototype.cli:cli'
